The Gray Lab has undertaken an effort to remove the compiler warnings from Rosetta.  All of them.  We are using Clang 3.1 with the -Wall flag to identify warnings and are focusing on cleaning up one library at a time.  This commit resolves the compiler warnings in the following libraries:
1.	ObjexxFCL
2.	basic
3.	numeric
4.	utility
5.	core.1
6.	core.2
7.	core.3
8.	core.4
9.	core.5

We are not touching external or boost.

A big thanks goes to Andrew for cleaning up ~100 unused parameter warnings!  And to Matt for participating as well.

Let's all try to keep these libraries free of warnings - warnings are pretty much of Sauron.

~ The Fellowship of the waRnINGs
	Brian Weitzner
	Mike Pacella
	Labonte

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@51625 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: f803c5421d307bb95c44c03b5ba6d8118a30be31

+++ rosetta_source/src/core/kinematics/FoldTree.cc
@@ -1954,10 +1954,10 @@ void
 FoldTree::setup_edge_counts() const
 {
 	// redimension?
-	if ( edge_count.size() != nres_ ) {
+	if ( (int)edge_count.size() != nres_ ) {
 		edge_count = utility::vector1<int>(nres_, 0);
 	}
-	if ( jump_edge_count.size() != num_jump_ ) {
+	if ( (int)jump_edge_count.size() != num_jump_ ) {
 		jump_edge_count = utility::vector1<int>(num_jump_, -1);
 	}
 
@@ -2243,10 +2243,10 @@ void
 FoldTree::update_jump_points() const
 {
 	// re-dimension?
-	if ( is_jump_point_.size() != nres_ ) {
+	if ( (int)is_jump_point_.size() != nres_ ) {
 		is_jump_point_ = utility::vector1<bool>(nres_, false);
 	}
-	if ( jump_point_.size() != num_jump_ ) {
+	if ( (int)jump_point_.size() != num_jump_ ) {
 		jump_point_ = utility::vector1<std::pair<int,int> >(num_jump_);
 	}
 
@@ -2284,7 +2284,7 @@ FoldTree::update_jump_points() const
 void
 FoldTree::update_jump_edge() const
 {
-	if ( jump_edge_.size() != num_jump_ ) {
+	if ( (int)jump_edge_.size() != num_jump_ ) {
 		jump_edge_ = utility::vector1<int>(num_jump_, 0);
 	}
 	int jump_index(0);
